Small optimization by replacing all uses of int with int8_t/uint8_t

+++ constants.h
@@ -1,8 +1,10 @@
 #ifndef _THEKBD_CONSTANT_H_
 #define _THEKBD_CONSTANT_H_
 
-const int countRows=7;
-const int countColumns=7;
+#include <stdint.h>
+
+const int8_t countRows=7;
+const int8_t countColumns=7;
 
 enum {
     keysTotalNumber=countRows*2*countColumns
